MB-15009: 1/3 Make frequency counter 8-bit
The frequency counter value is clamped to 8-bits but the supporting
code deals in 16-bit types. To make space for a defragmentation age
in StoredValue we update the code to properly manage the frequency
counter as 8-bits and reserve the other 8-bits in the tag for
future use (sv defrag visit age)

Reviewed-on: http://review.couchbase.org/106448
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/item.cc
@@ -368,12 +368,12 @@ item_info to_item_info(const ItemMetaData& itemMeta,
     return info;
 }
 
-void Item::setFreqCounterValue(uint16_t newValue) {
+void Item::setFreqCounterValue(uint8_t newValue) {
     auto taggedPtr = value.get();
     taggedPtr.setTag(newValue);
     value.reset(taggedPtr);
 }
 
-uint16_t Item::getFreqCounterValue() const {
+uint8_t Item::getFreqCounterValue() const {
     return value.get().getTag();
 }
